+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - set `All` constructor as private
 - set `Any` constructor as private
 - added `MapError` validator
+- added `HasNotKey` validator
 
 ## [0.1.0]